Added express router for managing routes
